+++ packages/autorest.python/tasks.py
@@ -13,6 +13,8 @@ from colorama import init, Fore
 from invoke import task, run
 import shutil
 import re
+from subprocess import check_call
+from pathlib import Path
 
 #######################################################
 # Working around for issue https://github.com/pyinvoke/invoke/issues/833 in python3.11
@@ -557,6 +559,8 @@ def regenerate_multiapi(c, debug=False, swagger_name="test"):
         "test/multiapi/specification/multiapisecurity/README.md",
         # create multiapi client with keyword only params
         "test/multiapi/specification/multiapikeywordonly/README.md",
+        # create basic multiapi client (package-name=multapicombiner)
+        "test/multiapi/specification/multiapicombiner/README.md",
     ]
 
     cmds = [_multiapi_command_line(
@@ -564,6 +568,7 @@ def regenerate_multiapi(c, debug=False, swagger_name="test"):
 
     _run_autorest(cmds, debug)
 
+    check_call(f"python {Path('test/multiapi/run_multiapi_combiner.py')} multiapicombiner", shell=True)
 
 @task
 def regenerate_package_mode(c, debug=False, swagger_group=None):