+++ packages/autorest.python/test/multiapi/Expected/AcceptanceTests/MultiapiCustomBaseUrl/multiapicustombaseurl/v1/aio/operations/_multiapi_custom_base_url_service_client_operations.py
@@ -32,6 +32,12 @@ ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T
 
 
 class MultiapiCustomBaseUrlServiceClientOperationsMixin(MultiapiCustomBaseUrlServiceClientMixinABC):
+    def _api_version(self, op_name: str) -> str:  # pylint: disable=unused-argument
+        try:
+            return self._config.api_version
+        except:  # pylint: disable=bare-except
+            return ""
+
     @distributed_trace_async
     async def test(self, id: int, **kwargs: Any) -> None:  # pylint: disable=inconsistent-return-statements
         """Should be a mixin operation. Put in 1 for the required parameter and have the correct api
@@ -55,7 +61,7 @@ class MultiapiCustomBaseUrlServiceClientOperationsMixin(MultiapiCustomBaseUrlSer
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version: str = kwargs.pop("api_version", _params.pop("api-version", "1.0.0"))
+        api_version: str = kwargs.pop("api_version", _params.pop("api-version", self._api_version("test") or "1.0.0"))
         cls: ClsType[None] = kwargs.pop("cls", None)
 
         request = build_test_request(