+++ packages/autorest.python/test/multiapi/Expected/AcceptanceTests/MultiapiWithSubmodule/multiapiwithsubmodule/submodule/v2/aio/operations/_multiapi_service_client_operations.py
@@ -33,6 +33,12 @@ ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T
 
 
 class MultiapiServiceClientOperationsMixin(MultiapiServiceClientMixinABC):
+    def _api_version(self, op_name: str) -> str:  # pylint: disable=unused-argument
+        try:
+            return self._config.api_version
+        except:  # pylint: disable=bare-except
+            return ""
+
     @distributed_trace_async
     async def test_one(self, id: int, message: Optional[str] = None, **kwargs: Any) -> _models.ModelTwo:
         """TestOne should be in an SecondVersionOperationsMixin. Returns ModelTwo.
@@ -57,7 +63,9 @@ class MultiapiServiceClientOperationsMixin(MultiapiServiceClientMixinABC):
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version: str = kwargs.pop("api_version", _params.pop("api-version", "2.0.0"))
+        api_version: str = kwargs.pop(
+            "api_version", _params.pop("api-version", self._api_version("test_one") or "2.0.0")
+        )
         cls: ClsType[_models.ModelTwo] = kwargs.pop("cls", None)
 
         request = build_test_one_request(
@@ -118,7 +126,9 @@ class MultiapiServiceClientOperationsMixin(MultiapiServiceClientMixinABC):
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version: str = kwargs.pop("api_version", _params.pop("api-version", "2.0.0"))
+        api_version: str = kwargs.pop(
+            "api_version", _params.pop("api-version", self._api_version("test_different_calls") or "2.0.0")
+        )
         cls: ClsType[None] = kwargs.pop("cls", None)
 
         request = build_test_different_calls_request(